Fix memory calculation when < 1% of memory is free
When memory dipped below 1% free, `check-memory-percent.sh` returned "MEM UNKNOWN" instead of "CRITICAL".

The reason is because `bc` returns floating point values without a leading zero by default.

Example:
```
[bbuckley@absinthe ~]# echo "scale=3; 1 / 1000 * 100" | bc -l| cut -d "." -f1

[bbuckley@absinthe ~]#
```

To fix this, I've substituted `bc` for `awk`.

This also makes the check work out-of-the-box on CentOS Minimal (`bc` is not included by default, but `awk` is).